% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaplot.R
\name{iiaplot_1}
\alias{iiaplot_1}
\title{Illustrated ANOVA}
\usage{
iiaplot_1(fixef = list(b_0 = 8, b_1 = -1, b_2 = 2), groups = c("A", "B",
  "C"), label = c("group", "outcome"))
}
\arguments{
\item{fixef}{list of three coefficients b_0, b_1, b_3, see Details}

\item{groups}{a vector of three strings, like c("A", "B", "C)}

\item{label}{vector of strings ("group", "outcome")

This function plots a diagram for an ANOVA model with
three groups.
This model has the following parameters:

b_0 is the intercept, the expected value of the reference group,
i.e. the first element in argument groups ("A").
b_1 and B_2 are the mean differences

The parameters can also be given as an unnamed list, as long as the elements comply to the order.

The line plot is augmented with illustrations, showing the two group effects
b_0 and b_1 as vertical bars and the slope parameters as triangles.}
}
\description{
plots a diagram for a linear model with one dichtotomous group and a metric
predictor
}
\examples{

iiaplot_1(fixef(b_0 = -2, b_1 = 1, b_2 = 1.5, b_3 = 1))


}
\author{
Martin Schmettow
}

