% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaplot.R
\name{iiaplot_2}
\alias{iiaplot_2}
\title{Illustrated regression line}
\usage{
iiaplot_2(fixef = list(b_0 = 0.5, b_2 = 1), range = c(0, 10),
  label = c("covariate", "outcome"))
}
\arguments{
\item{fixef}{list of four coefficients b_0 and b_2, see Details}

\item{range}{a vector of two values c(xmin, xmax)}

\item{label}{vector of strings ("covariate", "outcome")

This function plots a standard interaction diagram for a model with
one covariate and two groups.
This model has the following parameters:

b_0 is the intercept
b_2 is the slope of reference group

The parameters can also be given as an unnamed list, as long as the elements comply to the order.

The line plot is augmented with illustrations, showing the intercept as vertical bar and
slope parameters as triangles.}

\item{groups}{a vector of two strings, like c("A", "B")}
}
\description{
plots a diagram for a linear model with one dichtotomous group and a metric
predictor
}
\examples{

iiaplot_4(fixef(b_0 = -2, b_2 = 1.5))


}
\author{
Martin Schmettow
}

