% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betaplot.R
\name{iiaplot_1}
\alias{iiaplot_1}
\title{Illustrated interaction plot 1}
\usage{
iiaplot_1(fixef = c(b_0 = 0, b_1 = -8, b_2 = 1, b_3 = 2), range = c(0, 10),
  groups = c("A", "B"))
}
\arguments{
\item{fixef}{vector of four coefficients b_0 - b_3, see Details}

\item{range}{a vector of two values c(xmin, xmax)}

\item{groups}{a vector of two strings, like c("A", "B")

This function plots a standard interaction diagram for a model with one covariate and two groups. This model has the following parameters:

b_0 is the intercept, the expected value of the reference group, i.e. the first element in argument groups ("A").
b_1 is the grand difference towards the second group (b_2 = 0)
b_2 is the slope of reference group
b_3 is the interaction effect, which is the  difference in slope towards the reference group

The line plot is augmented with illustrations, showing the two group effects b_0 and b_1 as vertical bars and the slope parameters as triangles.}
}
\description{
plots a diagram for a linear model with one dichtotomous group and a metric
predictor
}
\examples{

iiaplot_1(fixef(b_0 = -2, b_1 = 1, b_2 = 1.5, b_3 = 1))


}
\author{
Martin Schmettow
}

